src/requirements.txt
Django>=1.11,<1.12
django-rest-swagger==2.1.2
djangorestframework==3.6.2
psycopg2==2.7.1
django-environ==0.4.3
gunicorn==19.7.1
boto3==1.4.4
jsonfield==2.0.1
django-formtools==2.0
django-exclusivebooleanfield==0.3.0

# nice-looking Bootstrappy forms
django-crispy-forms

# Sentry errors collection
# no version freeze yet - they broke it sometimes, by rarely
raven
